requirements.txt
SQLAlchemy>=1.4
psycopg-binary>=2.9
fastapi
uvicorn
celery
redis
alembic>=1.11.1
email-validator>=2.1.0.post1
httpx
PyJWT==2.*
passlib[bcrypt]==1.7.*
python-jose[cryptography]==3.*
bcrypt==3.2.2